more comments

+++ rust/lancedb/src/table.rs
@@ -1301,9 +1301,9 @@ impl NativeTable {
         ivf_params.sample_rate = index.sample_rate as usize;
         ivf_params.max_iters = index.max_iterations as usize;
         let hnsw_params = HnswBuildParams::default()
-            .num_edges(index.m)
-            .max_num_edges(index.m * 2)
-            .ef_construction(index.ef_construction);
+            .num_edges(index.m as usize)
+            .max_num_edges(index.m as usize * 2)
+            .ef_construction(index.ef_construction as usize);
         let pq_params = PQBuildParams {
             num_sub_vectors: num_sub_vectors as usize,
             ..Default::default()
